SDL: Float type support corrected
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@12871 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imtsdl/CSdlTools.cpp
@@ -177,7 +177,7 @@ QString CSdlTools::ConvertType(const QString& sdlTypeName, bool* isCustomPtr, bo
 			*isComplexPtr = false;
 		}
 
-		return QStringLiteral("float");
+		return QStringLiteral("double");
 	}
 
 	// A signed double-precision floating-point value
